+++ packages/core/src/js/index.js
@@ -18,5 +18,6 @@ export { installMulticombobox } from './multicombobox.js';
 export { installDrawer } from './drawer.js';
 export { installHovercard } from './hovercard.js';
 export { installToggleGroup } from './toggle-group.js';
+export { installContextMenu } from './context-menu.js';
 
 export const version = '0.0.0';